feat: Add linespacing to common text component
