+++ FlowUs/Localizable/de.lproj/Localizable.plist
@@ -2,6 +2,10 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+	<key>verify.description</key>
+	<string>Geben Sie die magischen Zahlen ein, die wir an Ihre E-Mail-Adresse geschickt haben, um Ihr Konto zu bestätigen, und genießen Sie die App</string>
+	<key>verify.header</key>
+	<string>Fast fertig</string>
 	<key>register.mail.verify</key>
 	<string>Ein Code zur Authentifizierung wird an die Email Adresse gesendet </string>
 	<key>register.country</key>
