+++ FlowUs/Views/shared/display/Typography.swift
@@ -65,14 +65,11 @@ struct CommonText: View {
     @Environment(\.colorScheme) var colorScheme
     var text = ""
     var semibold = false
-    var gutterBottom = true
     var body: some View {
         Text(.init(text))
             .font(Font.system(size: 16))
             .fontWeight(semibold ? .semibold : .regular)
-            .if(gutterBottom, transform: { view in
-                view.padding(.bottom, 4)
-            })
+            .lineSpacing(4)
     }
 }
